PDFBOX-2959: remove unneeded unboxing
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/trunk@1701472 13f79535-47bb-0310-9956-ffa450edef68

+++ pdfbox/src/main/java/org/apache/pdfbox/pdmodel/font/PDFont.java
@@ -227,7 +227,7 @@ public abstract class PDFont implements COSObjectable, PDFontLike
             int lastChar = dict.getInt(COSName.LAST_CHAR, -1);
             if (getWidths().size() > 0 && code >= firstChar && code <= lastChar)
             {
-                return getWidths().get(code - firstChar).floatValue();
+                return getWidths().get(code - firstChar);
             }
 
             PDFontDescriptor fd = getFontDescriptor();
